treewide: Make imports module-level granular
Set the rustfmt `imports_granularity` flag to `"Module"`, which makes
imports be separate items up to the last module component of the path.
This reformats a lot of code that was previously relying on the default
behavior of rustfmt of just preserving whatever style was written.

Since this is the first rustfmt configuration change to the entire tree,
this also removes the //noria/ and //msql-srv/ rustfmt.toml files, in
favor of just configuring at the root of the tree.

Change-Id: I5bade8b256edb3e15d2eb48ce9ed535b87473f25
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1851
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>