memory_state: Fallback between index types
If we miss when doing a lookup into a HashMap index in MemoryState, and
we have a BTreeMap index on the same columns, potentially use that index
to satisfy the lookup if it doesn't have a hole. This is one of the
"other solutions" to the problem that f4ba18c03 (dataflow: Use tag to
select index for self-lookups, 2022-02-01); except that unlike the
solution in that commit it doesn't also cause regular writes to
incorrectly miss, which is the reason it was reverted in
12f0df59e (Revert "dataflow: Use tag to select index for self-lookups",
2022-02-07). The original logictest that was added in 19d75f383 (server:
Thread through aggregates to post-lookup, 2022-01-28) but removed in
12f0df59e has been added back here, since it now works again, except
with the addition of a case covering normal writes too.

Change-Id: I388416472ca7b5ad10d0027a4d3d77186c8b2f15
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1855
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI