dataflow/grouped: properly handle deleting all rows
If we delete a row in a grouped operator such that we lose the ability
to construct state (such as if the extremum gets the current extreme
value deleted), we would previously return
`Err(ReadySetError::GroupedStateLost)`, a custom error variant that was
designed specifically for this case. The generic `Grouped` wrapper type
would then match on this error, and if it ever hit it do a lookup into
the parent, constructing the state using all rows for the group that
that returned. Only if the parent had *no rows* for the key (not a miss,
a filled hole with no rows), the internal grouped operator would return
`GroupedStateLost` again, and we'd just `?` it up to the top of the
domain until the entire domain failed.

Rather than repurposing the error type for "we have no values", then,
this commit changes the grouped operator's `apply` method to just return
`ReadySetResult<Option<DataType>>` instead, with `Ok(None)` indicating
that we lost state - then, if we reconstruct state based on the results
of a lookup from our parent and we return Ok(None), we substitute that
with either the `empty_value` from the operator, or `NULL`.

This fixes the case where we delete *all* the rows in a group for a
grouped operator like extremum that can lose state, which fixes 74
previously-failing logictests

Fixes: ENG-1032
Change-Id: Ibc7e167a8cc2136d37aa9c078a790aa49332d757
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1863
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>