demos: Solidus query set and schema dump.
This includes a subset of the queries from the solidus demo. These
queries are pulled directly from the app and should be parameterized in
the future to be used in benchmarks. The schema includes minor
modification to avoid current parse failures and key constraints that
we do not satisfy but is otherwise a mysqldump.

Change-Id: I39e16295009b677c397ae63feef776a562ca6195
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1860
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>