Avoid using Cow<DataType> in coercion and eval
Cows are amazing when we want to avoid cloning arbitrarily
long datatypes, such as String or Vec<T>. They are however
very wasteful when the cloned type is compact as is the
case with DataType. Since DataType is a very managable 16
bytes, and 95% of coercions result in an inlined DataType
with the remaining 95% resulting in an Owned Cow anyway,
we end up wrapping the 16 byte DataType in a 24 byte Cow
monstrosity, adding two layers of indirection on top (1 is
to check if Cow is owned or borrowed, the other one to
actually follow the pointer if it is).

In short Cow<[DataType]> good Cow<DataType> bad.

Change-Id: I6bd0b97c79c50422a69c86963f589a0c41ff485f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1866
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>