Pagination: Implement stub of Pagination NodeOperator
Implements a minimal stub of the Pagination NodeOperator. The operator
is based loosely on the TopK operator, because they have similar
requirements for ordering and self lookups. The LIMIT field is included
as a constant field in the operator, while the OFFSET field is intended
to be provided as a key value.

The topk::Order struct was pulled into a new ops::utils mod for reuse in
the Pagination operator.

The implementation of controller::schema::type_for_internal_column() was
updated with the assumption that the Pagination operator will mirror its
parents columns exactly, but append an additional column representing
the page number.

Change-Id: Ic11511b2a7e1573eb04b24fbef94cbf3e9f4425f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1870
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>