noria: Fix invalidate views and cache leader in `ControllerHandle`
We currently may issue an RPC on every leader RPC as we do not cache the
current leader in the ControllerHandle. This is expensive and
unnecessary, we can instead cache the leader address and only try to
retrieve a new leader if the RPC to the leader fails.

This performance hit was exacerbated by view invalidation occuring on
*any* error. This updates the view invalidation logic to only occur on
network errors as the view is still "technically" valid on many other
errors.

Change-Id: I0a17c4e641afb6c5ad8f95768b2053622f8b9bbd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1872
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>