Make read path fallible
Refactor the entire read path to be able to return a ReadySetError,
threading through (most notably) errors that occur in PostLookup
evaluation up through the various bits and pieces, terminating at making
the error type in `ReadReply::Normal` a ReadySetError instead of
just (). This allows us to actually return errors for invalid reads,
such as those that specify an invalid type or other value for eg a
PostLookupFilter.

Change-Id: I2fb045cb4285fe1c84929c8dc4e52d218914b769
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1878
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>