view: Just use a DataflowExpression for post-lookup filters
Rather than continuing to expand ViewQueryFilter to support all sorts of
different operators, which would potentially extend indefinitely up to
the point of completely becoming functionally identical to
DataflowExpression, just use a full-fledged
DataflowExpression (actually, Option<DataflowExpression>) as the type
for `ViewQuery::filter`. This comes at a slight performance
cost (likely, I haven't benchmarked it yet) as evaluating a
DataflowExpression requires recursively walking a nested data structure
currently (meaning we have to pay the cost of cache line misses and
loads), at the benefit of being able to support literally anything we
could do in a regular dataflow Filter node; if the performance cost
becomes significant, we can switch out our treewalking evaluator for
dataflow expressions for a bytecode interpreter pretty easily (this is
something I've been wanting to do for a while now regardless); I've also
had the thought that compiling dataflow expressions to WASM might be
cool.

In order to get this all to work, a couple of changes had to be made to
dataflow expressions:

- To allow `noria` to depend on `dataflow/expression` without a circular
  dep I had to move `like.rs` into `dataflow/expression`; this is
  probably where it should've been all along, as that's the only place
  it's actually used
- The type signature of `eval` had to be made slightly more polymorphic
  to admit either a `&[DataType]` or a `&[Cow<DataType>]`, this required
  adding some type hints to tests but overall should be zero-cost
  besides the type inference hit.

Change-Id: I21c16b5bc5dab5e79da65997226fd5b8e98c0127
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1882
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>