noria-client: Post-filter compound ranges
I don't know *why* I thought compound range queries (eg `x > ? AND y >
?`) could be implemented via a single lexicographic comparison on a
compound key for so long, but that's obviously not true - for example,
when comparing lexicographically, the compound key `(2, 1)` is *greater*
than the compound key `(1, 2)`, even though the second element is less.

There *may* be something clever we can do that takes advantage of
remapped upqueries to make this do the thing we want it to, but as it is
since there's no way to make a strict total out of piecewise comparison
of vectors (it violates the trichotomy law), and it's unsound to make
the keys in a rust BTreeMap compare in a way that isn't a strict total
order, the only actual recourse we have for compound ranges
unfortunately is to filter post-lookup.

Fixes: ENG-1011
Change-Id: I83b42a39feaae993d7582b0d41d2ca9409440e34
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1884
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>