ops/terraform: deploy eks cluster module and automation to build aws
The purpose of this CL is to get an EKS cluster deployed into the build
aws account. This CL builds upon previous CLs to fairly easily get a
mostly automated EKS cluster up and running.

A KMS key is deployed as part of this. Since it could be difficult to
switch things later, a CMK was provisioned and provided to EKS for
envelop KMS encryption of k8s secrets.

Terraform workspaces are used to split remote state. Whenever we create
a new environment, such as sandbox, we'll copy the tfvars/build-us-east-2
directory and adjust to fit the new environment.

Change-Id: Icf887b19f71290f6a046d7a50c43de2571c6b6e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1904
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>