solidus: Support generating data for solidus demo
This CL makes minor modifications to our generation utility to support
generating data for the solidus demo. It adds a new schema file
`solidus-generated` that can be used to append to the solidus database
with reasonable data.

It also adds 6 new solidus queries to the set of solidus queries in the
data.

Change-Id: I85728f7e4f14383d4707e38a6098c2debd44a37a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1928
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI