sql: Refactor subquery extraction pass to use Visitor
Now that we have a Visitor for select statements, we can gradually
refactor legacy code that was explicitly and *repeatedly* implementing
AST traversal to use that instead, making the code more pointed (as it
only contains specifically the bits that're relevant to that particular
pass) and more maintainable (as it becomes much more resilient to future
changes in the structure of our AST). I'm about to be making some
changes to this pass, so I figured now's as good a time as any to
refactor it to use a Visitor. Look at all that deleted code!!

Change-Id: I87d602afc5cdc20dac2ca7d5ae064bd733792d3e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1893
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI