sql: Refactor count_star_rewrite pass to use Visitor
This time not so much to simplify things (this actually adds a little
bit of code), but actually to fix a *bug* introduced by the constant
reimplementation of AST traversal - specifically, the previous version
of this pass only considered count(*) expressions at the top-level of
the FieldDefinitionExpression, not nested inside other AST nodes.
Switching to a Visitor gets us that fixed for free!

Change-Id: I0508ad856bb9b6a8565e728a02d2fde4b4fb3bfe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1896
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI