sql: Add support for uncorrelated EXISTS
Implement compilation of uncorrelated EXISTS subqueries, by:

- adding a `COUNT(*)` node after the subquery
- doing a join to the full subquery including the count
- adding a filter on the count to remove all rows where it is 0

In order to get the second step there to work properly, this also had to
add explicit support in mir->dataflow lowering for joins with no
conditions, which we implement here by projecting out a literal
bogokey-esque value on both sides of the join and using that as our join
key. From an *efficiency* perspective this is almost definitely pretty
cheap, as the count will only ever have one row per key - but there's
also a lot of room for MIR-level peephole optimization here, eg if we
discover a filter later on that "looks like" a join key (it filters on
one column from the left of the join and another from the right of the
join), we can teach the MIR optimizer to turn that filter into the join
key for the join! That will likely happen once we add support
for *correllated* EXISTS, which will require lifting correlated
conditions out of the EXISTS query using techniques similar to those
described in [0] and [1]

[0]: https://dl.acm.org/doi/10.1145/1247480.1247598
[1]: https://dl.acm.org/doi/10.1145/375663.375748

Change-Id: Iff0d90439a7587fe057819a88610f3661d9ac81e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1898
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>