Support additional CIDR in VPC supplemental
This adds the first condition to our parameters. This should allow our
existing work to deploy as is without requiring this string to be filled
out while allowing us to update those who need this.

This should be able to be applied over the top of an existing VPC
supplemental stack and will do manual testing of this.

Change-Id: I60088bfb4aad792a95c308d04b32a0041b354331
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1900
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>