ops/terraform: implement cross-account external-dns via roles
Building on some of the previous work in this chain of CLs, the
objective here is to get externalDNS assuming a role in another account
when it comes to managing our DNS zone records.

We've also opted to rename `readyset.name` to `build.readyset.name` to
make managing and searching DNS records at the VPN level a bit easier.

Each externalDNS deployment should be managing a single domain and zone,
so I've updated the deployment manifests and naming to reflect that
pattern.

The configuration within has been tested, and DNS records are now being
automatically provisioned in the new private hosted zone in Route53
under readyset.name.

Change-Id: Ide9eae1061c9bdd04042c2b09d2521587c5806df
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1905
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>