ops/terraform: add global terraform root-module for managed-services
This CL aims to provide a new Terraform root-module for AWS account-wide
resources to be deploy to, which inherently expand beyond a single
region's boundary.

Additionally, this CL adds a new private DNS zone and VPC zone
associations that are required for running our own dynamic k8s clusters.
Basically, if we want private DNS records that resolve over VPN and
in-AWS, this is the way to approach it.

It's worth noting that a "blackhole" vpc was created in the DNS
management account in order to break a circular dependency issue between
the private DNS zone resource, and the association/authorization. This
same VPC can be used in the future for any additional private zones we
may create for N domains.

Change-Id: Ib6e42318a791b6d2d8b97766adda036bb2c0df16
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1906
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>