readyset server/adapter: Log options at startup
For both the server and adapter, this will log the options struct at the
info level. This also adds a new `RedactedString` type, which wraps
regular strings, making it so when debug-printed that they will instead
show up only as "<anonymized>". A few items in the relevant options structs
have been converted from regular Strings to RedactedStrings, to help
avoid accidentally leaking credentials in our logs.

Change-Id: I396f8acc887064e246b73fbe5d129b05c915300a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1907
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>