ops/terraform: add terraform module for deploying kube-prometheus
This CL is intended to provide us with a succinct Terraform module for
deploying Prom/Grafana stacks.

I've deployed this module to the build account, and through the VPN, am
able to hit Prometheus and Grafana. For now, the metric retention
settings are set to 180d. We may need to raise or lower this value
depending on how well or poorly the disks handle the extra metrics.

The URL proposed for Grafana (service is up):
  - https://benchmark-grafana.build.readyset.name

The URL proposed for Prometheus (service is up):
  - https://benchmark-prometheus.build.readyset.name

Next up will be testing this w/ the benchmark binary and pushgateway.
@Justin may be able to kick the tires on the new setup to see if checks
the box for us.

Change-Id: I4fab90941c1e3bc83c9b7fd268aac4faf2615933
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1908
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>