Add noria error field to EXPLAIN LAST STATEMENT
Currently if the query goes to fallback there is no way
to retrive the noria error that caused it to go to
fallback. With this change a new field is added that
shows that field. In addition the destination values
are changed not to use noria and use ReadySet instead.

Change-Id: I8b765d323da97f73509942f681ec83e615defa82
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1909
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>