mir: split query installation into two parts
Currently, the only way to add a query to a graph via MIR involves a
single monolithic function that calculates reuse via a query graph, then
creates all the MIR nodes, then installs those MIR nodes as dataflow
nodes. This is problematic for subquery decorrelation, as we need the
ability to create a MIR graph for a SelectStatement that is
initially *invalid* (since it refers to columns in the outer query), and
then rewrite it via algebraic decorrelation to become valid as part of a
larger MIR graph.

In service of that, this commit splits out the actual creation of a
MirQuery from the process by which new queries are added to the graph,
so that we can use that directly when adding subqueries and delay the
actual creation of dataflow nodes until after we can run the
decorrelation pass.

Change-Id: I4b0b23e65d1f99227a633419ee52877925db7157
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1912
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>