Pagination: Implement Paginate in MIR
Replaces MirNodeInner::TopK with MirNodeInner::Paginate that is
converted into either a TopK or Paginate node in mir_to_flow. Using a
single node in MIR to represent both dataflow operators is done because:

- The construction and interfaces of the TopK and Paginate nodes are
  nearly identical, except that the Paginate node projects an additional
field (the page number) which is also part of its key.
- The TopK and Paginate nodes are mutually exclusive.

Two requirements for the Paginate node are enforced. First, the
offset parameter must be auto-parametrized by the adapter. Second,
Paginate nodes cannot be constructed for queries with range parameters.

Because the Paginate operator projects a page number as part of its key,
it should never need to project a bogokey. However, we may still need to
create a bogokey projection prior to the Paginate operator. In other
words, we may construct a paginate operator that orders the entire set
of records from a base table. However, the reader for that paginate node
will never request the entire set of records from the paginate node, as
it may only request one page at a time.

The page number field is added to the output of QueryGraph::view_key()
when a Paginate node is constructed. The page number field is not added
to QueryGraph.parameters, because these are parameters on query
relations, and the page number field would not fit nicely with that
paradigm.

The projected page number column is titled "___page_number___" to avoid
conflicts with customer column names. Before production, we may want to
either dynamically reassign this name if it conflicts with one in any of
the query's relations or map columns to their relatives in other nodes
instead of matching them by name.

Change-Id: I9ed1f67501ef7a84b0f16c1ba65b4fcba1773820
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1918
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>