server: Extract sql-passes to its own crate
Extract the noria_server::controller::sql::passes module to a new
`noria-sql-passes` crate, bringing along with it the `query_utils`
module (which is just in the `lib.rs` of the new crate) as it's needed
by some of the sql passes and makes sense to have alongside them anyway.

This is something I've been wanting to do for a while, not just because
the noria-server crate is *enormous* and *very* slow to compile, but
also because at some point in the future I'd like to refactor some of
noria-logictest to *use* these passes. I figured since I'm doing SQL
pass related stuff anyway I might as well get this over with.

Change-Id: Ia01050443e7c8fabb02d0a703b2b79322ab8cfd6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1914
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>