ops/terraform: create ips-refresher chart to refresh ECR creds
This CL is meant to solve for Kubelet (worker node) access to our
private ECR repositories. You usually don't want to grant such IAM
permissions at the entire host level, as this can raise some security
concerns.

Included in this CL is a Terraform module and Helm chart which will
ensure there is always a valid docker config file in the form of a k8s
secret in each k8s namespace, ready to be referenced by a deployment,
pod, job, etc. in the imagePullSecret parameter.

Change-Id: I5e31f15b74dd5e67fad8bf3a481aa09c111b9950
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1915
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>