ops/benchmarks: integrate ci benchmarks with k8s prom/push gateway
Now that we have Prometheus, Grafana, and Push Gateway setup in the
`build` Kubernetes cluster, we can wire our CI benchmarks up to them.
That's basically the goal of this CL. This should put us in a place
where we can start to compare benchmark results in Grafana, which seems
very exciting!

One thing I'd like to call out is that the `instance` label for each
benchmark pod is currently proposed as:

ci-benchmark-<benchmark pipeline build #>-<short_commit:0:7>

We can adjust this if needed, perhaps to just the commit sha, but it
really depends on how folks want to filter for one run vs another.

Here's a link to a Grafana dashboard I created to visualize what we're
seeing as a result of this CL:

https://bit.ly/3BjGt4W

Change-Id: I3e5531c811057763610768c80bb0f050d778884c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1917
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>