query_graph: Don't make a "computed_columns" relation for aggs
Rather than jamming the square peg of aggregated columns into the round
hole of query graph relations and edges, including hard-coding a
"computed_columns" string for the name of the relation everywhere, just
add `aggregates` and `group_by` fields to QueryGraph, using those to
do everything reuse-related that we were previously doing with the
"computed_columns" relation and the special `QueryGraphEdge::GroupBy`
edge type. This simplifies a lot of the code that had to special-case
handle these things, and also gets us one step closer to removing the
`function` attribute on `nom_sql::Column`.

Fixes: ENG-1074
Change-Id: Ic5d00cfe8394a0de03ae82545d88f4963ad7e23f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1922
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>