Don't drop secondary indices when resnapshotting
A snapshot currently drops all cfs, and then recreates the main index
only. This used to work just fine when we replayed the recipes one by
one, however since now the recipe is a monotonic constructs, the
indices are added before replication begins. We therefore now simply
keep all of the indices, and snapshot to all indices at the same time
compacting them all in the end.

Fixes: ENG-1030
Change-Id: I07e3f9b7a1ffcfc85ccc20e0e3f8c5203bd62180
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1910
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>