state: Fixed small bug on the memory state lookup code.
This commit fixes a small bug in the memory state lookup code, where we
mistakenly swapped the values of a tuple in an iteration with enumerate.
This was causing us to try to retrieve a value from the `Key` using the
actual column value instead of the position of said value in the columns
 list, making the method return `None` which then failed with an
 `expect` and made the domain to panic.

This commit also modifies the `shuffled_columns` test a little to make
such scenario works as expected.

Refs: ENG-1080
Change-Id: I9c0b87fb756326aec80bd7bca5e222d2ea5a046d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1921
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>