nom_sql,mir: Support ORDER BY on arbitrary expressions
Parse *any* arbitrary expression in the ORDER BY clause of a SELECT
query (in addition to just columns), and handle non-column expressions
in the order-by clause in MIR by first projecting them out in a separate
project node if necessary.

Along the way, this takes the opportunity to do a *little* bit of
cleanup of the surrounding code, including making the OrderType actually
optional in the AST and doing defaulting at interpretation time rather
than in the parser - since it's optional in the surface syntax, and
generally it's a better idea for ASTs to more closely mirror the surface
syntax of the language and leave semantics up to later passes.

Fixes: ENG-418
Change-Id: I5188ad441c32c1cd10e7d30931792c28a92d7b3b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1923
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>