sql-passes: refactor implied_tables pass to use a Visitor
In the continuing saga of refactoring various manual-traversal sql
passes to instead use a Visitor (cf 59cb47d5d (sql: Refactor
count_star_rewrite pass to use Visitor, 2022-02-10) and c9908ae74 (sql:
Refactor subquery extraction pass to use Visitor, 2022-02-10)), refactor
the expand_implied_tables pass to use a Visitor as well. This one fixes
a bug, since in *many* places we weren't properly traversing into
recursively nested subexpressions, in addition to generally making the
code a lot more terse and easier to maintain (In a separate change I
accidentally introduced a bug, as evidenced by the new test case that's
in here, and this new formulation of the pass makes that bug a lot more
obvious).

Change-Id: I9cda34f6d42ae2bf88d7a550154e3a6eee9df2da
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1930
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>