nom-sql: Replaced usage of `multispace` with new `whitespace` parser.
This commit replaces the usage of `multispace0` and `multispace1` with
`whitespace0` and `whitespace1` (respectively) in all of nom-sql, which
are parsers that can recognize SQL comments.
Some changes had to be made since `multispace` returned a single result
(the spaces consumed), whereas `whitespace` returns a vector with the
possible comments found.
So there were scenarios where we needed to have a parser that returned a
single value (that was going to be discarded), and that had to be
replaced with a hacky (but functional)
`map(whitespace, |_| "".as_bytes()`.

This also fixes an issue with `DROP TABLE` queries, where the binlog
replication was appending a multiline comment at the end of the
statement, like
```
DROP TABLE table_1; /* Generated by server */
```
This was causing Noria to throw an error, instead of handling the query.

Refs: ENG-1145
Change-Id: I5cf4be57f45f7c07753aa67ae93542a322350a58
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2060
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>