Remove appending db name
Previously we were automatically appending a default db name when
registering the replication url. This is bug prone as the database name
is usually already in the replication url, causing us to append the
default database name on top of the conn url.

This breaks postgres testing. It's utterly confusing how it doesn't
break for mysql logictests with fallback. Maybe the extra `/db-name` on
the end gets ignored by the mysql connector?

In either case, changing this does *not* break mysql logic tests run
with fallback, and enables us to test posgres logic tests with fallback.

Change-Id: I8577130c15463a1e05aca6a10f681ebcf1584849
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1932
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI