nom-sql: Subquery alias isn't actually optional
Previously, we were only optionally parsing aliases on joined-to
subqueries in nom-sql - this diverges from the behavior of postgresql
and mysql, both of which require an alias on all joined-to
subqueries (which is unsurprising, as otherwise there's no way to
disambiguate references to that subquery's columns!). Making this
non-optional at parse time and in the AST saves on a little extra work
unwrapping it later on down the line.

Change-Id: If9712b7df49c73750681adc786c1fe822c9963a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1935
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>