benchmarks: Better zipfian random number generation.
Updates the zipfian generator to use a random mapping from randomly
generated value to value in the range we are generating for. This
prevents all the randomly generated values from being in the same
contiguous range of values, i.e. 0 is the most popular, 1 is the second
most popular.

This requires modifies the parameter generation path to reuse a column
generator, thus enabling sharing the column generation object via a
lock.

Change-Id: I4096fabd862f4cf6207c8ae618524ca1aa87d2d9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1936
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>