Move float coercion to a separate module
* Moves float coercion to a separate module
* Implements missing coercions
* Fixes implementation to match MySQL coercions
* Fixed data leak in some conversion errors

REMARK: float to unsigned coercion and float to
unsigned cast behave differently in MySQL. Negative
values can't be coerced but can be casted explicitly.
This implementation implements the coercion not
the casting behaviour, amongst over things because
cast is only to u64, and coerce can be to any
sized int.

REMARK2: Our numeric type is not similar to the
SQL numeric/decimal types, and this CL does not
fix that. So all numeric/decimal conversions are
done as best effort, but there is hardly anything
to test there.

Change-Id: If331e267cdf2d54bd173078a3e3cb872275128f8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1943
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>