sql: Compile subqueries as part of the same MIR graph
Rather than compiling subqueries by separately compiling them into
anonymous views then syntactically rewriting them in the original query
into references to those views, compile non-CTE subqueries together with
the overall MIR graph for the query, including both on the right-hand
side of a join and on the right-hand side of an EXISTS expression (but
not in direct scalar expression position anymore; this was untested and
broken before, and we can bring it back later if we need to). This
preserves all original behavior, but unlocks MIR rewrite and
optimization passes that cross subquery boundaries, which decorrelation
will need to do.

Under the hood this requires tracking whether a relation represented by
a particular `QueryGraphNode` is a subquery, and if so recording both
the SelectStatement and the QueryGraph for that subquery. This is
a *little* annoying, since everywhere we pass a `QueryGraph` we
currently have to also pass the original SelectStatement; I've left a
TODO comment here to just make the original SelectStatement a field of
the QueryGraph (or rather anything we need out of it that isn't already
part of the QueryGraph).

CTEs are still done syntactically using the compile-to-a-view approach
for now; since they can't be correlated and we don't currently have
any *optimization* passes that cross the subquery boundary besides
decorrelation, this makes sense, but if that changes in the future it
would be pretty trivial to make them part of the same graph.

Fixes: ENG-1077
Change-Id: Icbd65c980146aea50cc6a9284ea2039a0e4ceb6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1937
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>