controller: Simplified `RecipeSpec`.
Modified the `require_leader_ready` field in `RecipeSpec` to be a
boolean instead of an `Option`.
The logic we previously had, is that if we had `None`, we would
interpret the flag as true.

This commit modifies the field to remove the `Option`, and correctly
implements the `Default` trait to reflect that behaviour, thus making it
 more readable

Refs: ENG-1082
Change-Id: I68f0ec2836a720b01fa34b6f88e46ee93948a14a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1938
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>