controller: Added capability to run a migration in dry-run mode.
This commit introduces the capability of performing an `extend_recipe`
in dry-run mode.
That means that the migration will not run the `MigrationPlan::apply`,
which is the one that sends the messages to the other workers.
For now, this capability won't be used, and the use will be added in
future commits.

This is not very clean, since _it could be possible_ to run
`extend_recipe` with the `dry_run` flag set, and then commit the
dataflow changes (so, no messages would be sent to the workers, but the
changes would be persisted anyways).

With the current state of the system, this is the easiest and cleanest
way (that I could think of at least) without incurring into a heavy
refactor or having separate methods that perform almost the same logic.

The idea with this is to add a separate endpoint in the controller to
handle dry-runs in a future commit. We want that for two reasons:
1. So we can avoid using the `extend_recipe` endpoint, which will block
the writer thread for a migration we already know we are not going to
persist.
2. To make sure we don't persist the dataflow state at the end (which is
 a given from 1, since the dataflow state reader can't be modified
 anyways).

Refs: ENG-1082
Change-Id: I3205421b3ebbe3cc3a34cf6d766f92a47e925af5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1939
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>