controller: Added dry-run endpoint.
This commit introduces a new endpoint, which accepts the same type of
requests as `install_recipe` or `extend_recipe`, but guarantees that the
 migration will be run in dry-mode.
This is a separate endpoint (instead of just enhancing the `RecipeSpec`
with a `dry_run` flag) because:
1. The `extend_recipe` endpoint would block the writer thread for a
migration we already know we are not going to persist.
2. We can guarantee we don't persist the dataflow state at the end
(which is a given from 1, since the dataflow state reader can't be
modified anyways).

Refs: ENG-1082
Change-Id: I1391bf01cc3cd9fa10094a8ec671fc8f7b02bf82
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1940
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>