sql-passes: Refactor alias-removal pass to use a Visitor
Another day, another bug caused by a manual traversal implemented in a
sql rewrite pass fixed by refactoring the whole pass to use a Visitor -
this time, the alias-removal pass wasn't properly traversing into
subqueries. It now does that, and also properly implements scope
shadowing for table aliases that may be correlated to outer queries.
Also, some vestigial security-universe-related stuff ("context") is
gone, since this was the only thing that was using it and it has no need
to.

Change-Id: I87e6eeafd5d18d50aa680c72d6cf04d79eccf382
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1941
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>