Remove logic tests w/ fallback
Running logic tests with fallback doesn't provide any value. What
happens is that the writes go to the upstream database (MySQL or
Postgres) and replication doesn't have time to catch up by the time we
execute our reads against the adapter. The result is that because Noria
doesn't know about the tables yet, the queries all go to fallback. This
results in us essentially just testing the upstream database more or
less directly.

Change-Id: I4f91167387d5c2caf8e35da182674b655b63defd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1942
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>