sql: Emit dependent joins for correlated subqueries
We need a targeted way to detect both which joins have potential
correlated nodes that need to be removed by rewrite passes, and also
within those rewrite passes to detect when we've sufficiently rewritten
a query. To this end, this introduces a new "DependentJoin" MIR node
type, inspired by the HyPer paper[0], which represents a regular inner
join, but allowing for nodes on the right side to reference columns on
the left side. During conversion of a query to MIR, we analyze subqueries
*syntactically* to determine if they're correlated, and use that
information to decide whether to emit inner or dependent joins (left
dependent joins error out for now; later on we can decide if it makes
sense to allow them). As this has no actual computational representation
in incremental dataflow, this errors out if we ever actually find it
when lowering to dataflow - the idea is that any rewrite pass will have
fully eliminated any dependent joins by the time we start converting to
dataflow, or otherwise we'll bail out and just send the query to
fallback.

[0]: https://bit.ly/3I5Gts0

Refs: ENG-1076
Change-Id: Ic787b4ab7b99019b39eae88726192c6125409322
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1944
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>